Suppress warnings and apply multiple bugfixes
Add /.build_ci to .gitignore and update CMake to suppress unused-parameter warnings for MSVC and non-MSVC builds; remove a unity-build preset entry. Fix TestRTS console issues (use printf-style AddLog and safe strcpy_s). Handle an invalid DebugShape type and tidy switch logic. Correct template/member lookup and sizing in ECS (ComponentAccessor call and static_cast for archetype list size). Extend JSON/LDS switch to cover more value types and replace direct base-class calls with this-> in LDSEnumFlagsPtr template wrappers. Update LDS object serializer call to match API (dot vs arrow). Remove stray BOM/#pragma once from Worlds.cpp. These changes reduce compiler warnings, fix unsafe/incorrect usages, and adapt to API/ABI changes.

+++ src/PhoenixSim/ECS/ArchetypeManager.h
@@ -89,7 +89,7 @@ namespace Phoenix
             ArchetypeManager(TAllocator& allocator, const Config& config)
                 : ComponentDefinitions(allocator, config.MaxComponentDefs)
                 , ArchetypeDefinitions(allocator, config.MaxArchetypeDefs)
-                , ArchetypeLists(allocator, { config.ArchetypeListSize + sizeof(TArchetypeList), config.MaxArchetypeLists })
+                , ArchetypeLists(allocator, { static_cast<uint32>(config.ArchetypeListSize + sizeof(TArchetypeList)), config.MaxArchetypeLists })
                 , EntityHandles(allocator, config.MaxEntities)
                 , Configuration(config)
             {
